+++ src/database/sql/User.model.sql
@@ -2,11 +2,11 @@
 -- but since I'm not using it yet we will write raw SQL code here.
 
 CREATE TABLE IF NOT EXISTS users (
-    id SERIAL PRIMARY KEY,
+    id uuid PRIMARY KEY DEFAULT gen_random_uuid(),
     email VARCHAR(100) UNIQUE NOT NULL,
     name VARCHAR(50) NOT NULL,
     password VARCHAR(255) NOT NULL,
-    role VARCHAR(7) NOT NULL,
+    role VARCHAR(7) CHECK (role IN ('patient', 'doctor', 'finance')) NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );